修复精度问题

+++ src/common/common.h
@@ -32,7 +32,7 @@ struct Value {
     ColType type;  // type of value
     union {
         int int_val;      // int value
-        float float_val;  // float value
+        double float_val;  // float value  --by 星穹铁道高手
     };
     std::string str_val;  // string value
 
@@ -43,7 +43,7 @@ struct Value {
         int_val = int_val_;
     }
 
-    void set_float(float float_val_) {
+    void set_float(double float_val_) {
         type = TYPE_FLOAT;
         float_val = float_val_;
     }
@@ -60,8 +60,8 @@ struct Value {
             assert(len == sizeof(int));
             *(int *)(raw->data) = int_val;
         } else if (type == TYPE_FLOAT) {
-            assert(len == sizeof(float));
-            *(float *)(raw->data) = float_val;
+            assert(len == sizeof(double));
+            *(double *)(raw->data) = float_val;
         } else if (type == TYPE_STRING) {
             if (len < (int)str_val.size()) {
                 throw StringOverflowError();
@@ -155,7 +155,7 @@ private:
                         return value;
                     }else if(type==TYPE_FLOAT){
                         char* charPointer2 = reinterpret_cast<char*>(record.data + meta.offset);  
-                        float float_val = *reinterpret_cast<float*>(charPointer2);
+                        double float_val = *reinterpret_cast<double*>(charPointer2);
                         value.set_float(float_val);
                         value.init_raw(meta.len);
                         return value;
